--- app/src/main/java/com/github/saikcaskey/pokertracker/data/utils/defaultDispatchersProviders.kt
@@ -1,11 +1,0 @@
-package com.github.saikcaskey.pokertracker.data.utils
-
-import com.github.saikcaskey.pokertracker.domain.CoroutineDispatchers
-import kotlinx.coroutines.Dispatchers
-
-val defaultCoroutineDispatchersProviders = object : CoroutineDispatchers {
-    override val main = Dispatchers.Main
-    override val io = Dispatchers.IO
-    override val default = Dispatchers.Default
-    override val unconfined = Dispatchers.Unconfined
-}
