+++ app/src/main/java/com/github/saikcaskey/pokertracker/di/DispatchersProvidersModule.kt
@@ -1,6 +1,6 @@
 package com.github.saikcaskey.pokertracker.di
 
-import com.github.saikcaskey.pokertracker.data.utils.defaultCoroutineDispatchersProviders
+import com.github.saikcaskey.data.utils.defaultCoroutineDispatchersProviders
 import com.github.saikcaskey.pokertracker.domain.CoroutineDispatchers
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.get
