+++ app/src/main/java/com/github/saikcaskey/pokertracker/presentation/root/DefaultRootComponent.kt
@@ -9,7 +9,7 @@ import com.arkivanov.decompose.router.stack.pop
 import com.arkivanov.decompose.router.stack.popTo
 import com.arkivanov.decompose.router.stack.push
 import com.arkivanov.decompose.value.Value
-import com.github.saikcaskey.pokertracker.data.utils.defaultCoroutineDispatchersProviders
+import com.github.saikcaskey.data.utils.defaultCoroutineDispatchersProviders
 import com.github.saikcaskey.pokertracker.domain.util.nowAsLocalDateTime
 import com.github.saikcaskey.pokertracker.domain.CoroutineDispatchers
 import com.github.saikcaskey.pokertracker.domain.components.DayDetailComponent
